Extract upload logic to dedicated component
